refactor: 사용하지 않는 import문/상수 삭제, 비지니스 로직 서비스로 리팩토링, 매직스트링 상수로 추출
